Put accessed boards to local storage
