+++ frontend/src/components/board/CreateColumnButton.tsx
@@ -1,7 +1,7 @@
 import ColumnCreationForm from "@components/board/ColumnCreationForm"
 import { Dialog, DialogContent, IconButton, Tooltip } from "@mui/material"
 import Box from "@mui/material/Box"
-import { getId } from "@services/Utils"
+import { getId } from "@/services/utils"
 import { useState } from "react"
 
 import { useAddColumnMutation } from "@/state/apiSlice"
